Fixed the profile-fields admin page
